IE10,11 would sometimes throw WrongDocumentError when appending content-nodes to fragments in Hazard processing. The offending content-nodes were in a document with URL of 'about:blank' (although this didn't always cause an error).
To work-around this began to use cloneDocument() in cacheAdd()/cacheLookup(),
and cloneDocument() calls createDocument(srcDoc) which attempts to create
a document with the same URL as srcDoc.

This does seem to address the problem in IE, although createDocument(srcDoc)
doesn't always create a document with correct URL, e.g. old Safari.

TODO: needs more cross-browser testing.