Add checkElementPerformance() to check Hazard templates for IE11 perf issues: - <element style="..."> - <li value="..."> - <custom-element>
The custom-element check requires filtering to ignore haz:* elements and
hf:* (or hf-) elements.

The hf: namespace required access to the frameset definition from the Processor definitions, which required the frameset definition to be passed during Processor instantiation.

A ref to the frameset definition wasn't available until after the frameset definition was initialized (which included processor init too).

Frameset definition is now split into two parts, the second of which is now
HFramesetDefinition.prototype.preprocess(). This is where processors are instantiated, and now there is already a ref to the frameset definition.